Music and Skybox
Edited skybox to be nighttime so it is easier on the eyes.

Edited music playlist script so that it is actually functional and renamed some songs in the audio folder as a test. There are currently 2 copies of all audio files. 1 copy has the index number from which the music playlist will call. The other is the name of the song. When all songs are added, an array with all the names will be made to lessen space.